+++ lib/hooks/useCookies/useCookies.d.ts
@@ -1,0 +1,2 @@
+import { Cookies } from '../../action';
+export declare function useCookies<T extends Cookies>(): T;