+++ lib/hooks/useSearch/useSearch.d.ts
@@ -1,0 +1,1 @@
+export declare function useSearch<D>(): D;