+++ src/hooks/useBody/useBody.ts
@@ -1,0 +1,5 @@
+import { useAction } from '../useAction'
+
+export function useBody<T> (): T {
+  return useAction().body as T
+}