+++ src/hooks/useCookies/useCookies.ts
@@ -1,0 +1,7 @@
+import { useAction } from '../useAction'
+
+import { Cookies } from '../../action'
+
+export function useCookies<T extends Cookies> (): T {
+  return useAction().cookies as T
+}