+++ src/hooks/useFiles/useFiles.ts
@@ -1,0 +1,5 @@
+import { useAction } from '../useAction'
+
+export function useFiles<T> (): T {
+  return useAction().files as T
+}