+++ src/hooks/useSearch/useSearch.ts
@@ -1,0 +1,5 @@
+import { useAction } from '../useAction'
+
+export function useSearch <D> (): D {
+  return useAction().search as unknown as D
+}